refactored image task to directly write files to variation folder
