+++ gulp/tasks/copy.js
@@ -5,27 +5,6 @@ const watch = require('gulp-watch');
 const runSequence = require('run-sequence');
 const config = require('./../config');
 
-// copy resources folders
-function moveRessources(baseDir) {
-
-    const mergedStream = merge();
-    const modulePaths = globule.find(baseDir + '/*');
-
-    modulePaths.forEach((moduleDir) => {
-        const stream = gulp.src(moduleDir + '/resources/**');
-
-        globule.find([
-            moduleDir + '/*',
-            '!' + moduleDir + '/resources'
-        ]).forEach((path) => {
-            stream.pipe(gulp.dest(path + '/resources/'));
-        });
-
-        mergedStream.add(stream)
-    });
-
-    return mergedStream;
-}
 
 gulp.task('copy:js:dev', function () {
     return gulp.src(config.srcDir + '/components/*/resources/js/**/*.js')
@@ -50,18 +29,6 @@ gulp.task('copy:layouts:dev', function () {
 
     return mergedStream;
 });
-
-
-// @TODO Remove old resources folders before copying
-gulp.task('move:resources:dev', function () {
-    return moveRessources(config.devDir);
-});
-
-gulp.task('move:resources:dist', function () {
-    return moveRessources(config.distDir);
-});
-
-
 
 gulp.task('watch:js', function () {
     watch(config.srcDir + '/components/*/resources/js/**/*.js', config.watch, function () {
